+++ src/lib.rs
@@ -50,7 +50,6 @@ DOM structure itself based on: https://github.com/SimonSapin/rust-forest/tree/ma
 */
 
 #![warn(missing_docs)]
-#![deny(unused_import_braces)]
 
 #[macro_use]
 extern crate svgparser;